# artifacts/telegram-bot/src/api.ts
/**
 * Typed API client — calls the local api-server at localhost:8080.
 * All key service operations go through REST to keep logic in one place.
 */

const BASE = process.env.API_BASE_URL ?? "http://localhost:8080";

async function post<T>(path: string, body: unknown): Promise<T> {
  const res = await fetch(`${BASE}${path}`, {
    method: "POST",
    headers: { "Content-Type": "application/json" },
    body: JSON.stringify(body),
  });
  const data = await res.json() as T;
  return data;
}

async function get<T>(path: string): Promise<T> {
  const res = await fetch(`${BASE}${path}`);
  return res.json() as Promise<T>;
}

// ─── Prices ───────────────────────────────────────────────────────────────────

export interface PricesResponse {
  basicPrice: number;
  proPrice: number;
  basicPriceFormatted: string;
  proPriceFormatted: string;
  paymentEnabled?: boolean;
  bank?: { name: string; bin: string; account: string; holder: string };
}

export interface OrderResponse {
  orderId: number;
  orderCode: string;
  amount: number;
  amountFormatted: string;
  expiresAt: string;
  qrUrl: string;
  bank: { name: string; account: string; holder: string };
  error?: string;
}

export async function createOrder(telegramId: string, plan: "basic" | "pro", username?: string): Promise<OrderResponse> {
  return post<OrderResponse>("/api/payment/orders", { telegramId, plan, username });
}

let _pricesCache: PricesResponse | null = null;
let _pricesCacheAt = 0;
const PRICES_TTL = 5 * 60 * 1000; // 5 minutes

export async function getPrices(): Promise<PricesResponse> {
  if (_pricesCache && Date.now() - _pricesCacheAt < PRICES_TTL) return _pricesCache;
  try {
    const data = await get<PricesResponse>("/api/prices");
    _pricesCache = data;
    _pricesCacheAt = Date.now();
    return data;
  } catch {
    return _pricesCache ?? { basicPrice: 20000, proPrice: 99000, basicPriceFormatted: "20.000đ", proPriceFormatted: "99.000đ" };
  }
}

// ─── Trial ────────────────────────────────────────────────────────────────────

export interface TrialStatus {
  hasTrialLeft: boolean;
  remaining: number;
  total: number;
}

export async function checkTrial(telegramId: string): Promise<TrialStatus> {
  return post("/api/keys/trial/check", { telegramId });
}

export async function useTrial(telegramId: string): Promise<{ allowed: boolean; usedCount: number; remaining: number; reason?: string }> {
  return post("/api/keys/trial/use", { telegramId });
}

// ─── Keys ─────────────────────────────────────────────────────────────────────

export interface ValidateResponse {
  valid: boolean;
  reason: string;
  expiresAt?: string;
  dailyUsesLeft?: number;
  totalUsesLeft?: number;
  retryAfter?: string;
  maxConcurrent?: number;
}

export async function validateKey(rawKey: string, telegramId?: string): Promise<ValidateResponse> {
  const qs = new URLSearchParams({ key: rawKey });
  if (telegramId) qs.set("telegram_id", telegramId);
  return get(`/api/keys/validate?${qs}`);
}

export interface ActivateResponse {
  success: boolean;
  reason?: string;
  expiresAt?: string;
  maxConcurrent?: number;
  dailyLimit?: number;
  maxTotalUses?: number;
}

export async function activateKey(opts: {
  key: string;
  telegramId: string;
  username?: string;
  firstName?: string;
}): Promise<ActivateResponse> {
  return post("/api/keys/activate", opts);
}

export interface UseKeyResponse {
  allowed: boolean;
  keyId?: number;
  reason?: string;
  retryAfter?: string;
}

export async function useKey(rawKey: string, telegramId: string): Promise<UseKeyResponse> {
  return post("/api/keys/use", { key: rawKey, telegramId });
}

export async function releaseKey(keyId: number): Promise<void> {
  await post("/api/keys/release", { keyId });
}

// ─── Check ────────────────────────────────────────────────────────────────────

export interface CheckResult {
  input: string;
  email: string | null;
  status: "live" | "die" | "deactivated" | "locked" | "error";
  user: string | null;
  plan: string | null;
  error: string | null;
}

/**
 * Check a single account via the SSE /api/check endpoint — same code path as the web checker.
 * Falls back with one retry on transient network_error.
 */
export async function checkSingle(mode: "account" | "session", line: string): Promise<CheckResult> {
  async function attempt(): Promise<CheckResult | null> {
    return new Promise((resolve) => {
      let result: CheckResult | null = null;
      checkBulk({
        mode,
        rawText: line,
        concurrency: 1,
        onResult: (r) => { result = r; },
        onDone: () => resolve(result),
        onError: () => resolve(null),
      });
    });
  }

  const first = await attempt();
  if (first && first.status !== "error") return first;
  // If null (stream error) or transient network/error status — wait 2 s then retry once
  if (!first || first.error === "network_error" || first.status === "error") {
    await new Promise<void>((r) => setTimeout(r, 2000));
    const second = await attempt();
    if (second) return second;
  }
  // Return whatever we have, or a generic error result
  return first ?? {
    input: line,
    email: null,
    status: "error",
    user: null,
    plan: null,
    error: "network_error",
  };
}

/** Streams bulk check results via SSE, calling onResult for each account. */
export async function checkBulk(opts: {
  mode: "account" | "session";
  rawText: string;
  concurrency?: number;
  proxies?: string[];
  onResult: (result: CheckResult & { index: number; completed: number }) => void;
  onDone: (total: number) => void;
  onError?: (err: Error) => void;
}): Promise<void> {
  try {
    const res = await fetch(`${BASE}/api/check`, {
      method: "POST",
      headers: { "Content-Type": "application/json" },
      body: JSON.stringify({
        mode: opts.mode,
        rawText: opts.rawText,
        concurrency: opts.concurrency ?? 2,
        proxies: opts.proxies ?? [],
      }),
    });

    if (!res.body) throw new Error("No SSE body");
    const reader = res.body.getReader();
    const dec = new TextDecoder();
    let buf = "";

    while (true) {
      const { value, done } = await reader.read();
      if (done) break;
      buf += dec.decode(value, { stream: true });
      const events = buf.split("\n\n");
      buf = events.pop() ?? "";
      for (const block of events) {
        for (const line of block.split("\n")) {
          if (!line.startsWith("data: ")) continue;
          try {
            const ev = JSON.parse(line.slice(6));
            if (ev.type === "result") opts.onResult(ev.data);
            else if (ev.type === "done") opts.onDone(ev.total);
          } catch { /* ignore parse errors */ }
        }
      }
    }
  } catch (e) {
    opts.onError?.(e as Error);
  }
}