# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChatGPT Account Checker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payment/orders:
    post:
      operationId: createOrder
      tags: [public]
      summary: Create a payment order (called by bot)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "200":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
...
components:
  schemas:
...
    CreateOrderInput:
      type: object
      required: [telegramId, plan]
      properties:
        telegramId: { type: string }
        username: { type: string }
        plan:
          type: string
          enum: [basic, pro]
    OrderResponse:
      type: object
      properties:
        orderId: { type: number }
        orderCode: { type: string }
        amount: { type: number }
        amountFormatted: { type: string }
        expiresAt: { type: string }
        qrUrl: { type: string }
        bank:
          type: object
          properties:
            name: { type: string }
            account: { type: string }
            holder: { type: string }
...