# lib/api-zod/src/generated/types/createOrderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ChatGPT Account Checker API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderInputPlan } from './createOrderInputPlan';

export interface CreateOrderInput {
  telegramId: string;
  username?: string;
  plan: CreateOrderInputPlan;
}