# lib/api-zod/src/generated/types/orderResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ChatGPT Account Checker API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderResponseBank } from './orderResponseBank';

export interface OrderResponse {
  orderId?: number;
  orderCode?: string;
  amount?: number;
  amountFormatted?: string;
  expiresAt?: string;
  qrUrl?: string;
  bank?: OrderResponseBank;
}